Organize package into subpackages
