+++ smtpburst/__init__.py
@@ -6,14 +6,10 @@ from . import (
     cli,
     datagen,
     attacks,
-    report,
     discovery,
-    nettests,
+    reporting,
     inbox,
-    tls_probe,
-    ssl_probe,
     pipeline,
-    rdns,
     proxy,
 )
 
@@ -23,13 +19,9 @@ __all__ = [
     "cli",
     "datagen",
     "attacks",
-    "report",
     "discovery",
-    "nettests",
+    "reporting",
     "inbox",
-    "tls_probe",
-    "ssl_probe",
     "pipeline",
-    "rdns",
     "proxy",
 ]
